+++ capsicum/src/fcntl.rs
@@ -32,8 +32,8 @@ pub enum Fcntl {
 /// ```
 /// # use capsicum::{Fcntl, FcntlsBuilder};
 /// let rights = FcntlsBuilder::new()
-///     .add(Fcntl::GetFL)
-///     .add(Fcntl::SetFL)
+///     .allow(Fcntl::GetFL)
+///     .allow(Fcntl::SetFL)
 ///     .finalize();
 /// ```
 #[derive(Clone, Debug, Default)]
@@ -44,6 +44,11 @@ impl FcntlsBuilder {
         FcntlsBuilder::default()
     }
 
+    #[deprecated(since = "0.4.0", note = "use FcntlsBuilder::allow instead")]
+    pub fn add(&mut self, right: Fcntl) -> &mut FcntlsBuilder {
+        self.allow(right)
+    }
+
     /// Allow an additional fcntl
     ///
     /// # Examples
@@ -51,9 +56,9 @@ impl FcntlsBuilder {
     /// # use capsicum::{Fcntl, FcntlsBuilder};
     ///
     /// let mut builder = FcntlsBuilder::new();
-    /// builder.add(Fcntl::GetFL);
+    /// builder.allow(Fcntl::GetFL);
     /// ```
-    pub fn add(&mut self, right: Fcntl) -> &mut FcntlsBuilder {
+    pub fn allow(&mut self, right: Fcntl) -> &mut FcntlsBuilder {
         self.0 |= right as u32;
         self
     }
@@ -62,9 +67,17 @@ impl FcntlsBuilder {
         FcntlRights(self.0)
     }
 
-    #[deprecated(since = "0.4.0", note = "If you still need this method, please file an issue at https://github.com/dlrobertson/capsicum-rs/issues")]
+    #[deprecated(
+        since = "0.4.0",
+        note = "If you still need this method, please file an issue at https://github.com/dlrobertson/capsicum-rs/issues"
+    )]
     pub fn raw(&self) -> u32 {
         self.0
+    }
+
+    #[deprecated(since = "0.4.0", note = "use FcntlsBuilder::deny instead")]
+    pub fn remove(&mut self, right: Fcntl) -> &mut FcntlsBuilder {
+        self.deny(right)
     }
 
     /// Remove an allowed fcntl from the builder's list.
@@ -73,12 +86,12 @@ impl FcntlsBuilder {
     /// ```
     /// # use capsicum::{Fcntl, FcntlsBuilder};
     /// let mut common_builder = FcntlsBuilder::new();
-    /// common_builder.add(Fcntl::GetFL);
-    /// common_builder.add(Fcntl::SetFL);
+    /// common_builder.allow(Fcntl::GetFL);
+    /// common_builder.allow(Fcntl::SetFL);
     /// let mut restricted_builder = common_builder.clone();
-    /// restricted_builder.remove(Fcntl::SetFL);
+    /// restricted_builder.deny(Fcntl::SetFL);
     /// ```
-    pub fn remove(&mut self, right: Fcntl) -> &mut FcntlsBuilder {
+    pub fn deny(&mut self, right: Fcntl) -> &mut FcntlsBuilder {
         self.0 &= !(right as u32);
         self
     }
@@ -100,7 +113,7 @@ impl FcntlsBuilder {
 /// use nix::fcntl::{FcntlArg, OFlag, fcntl};
 /// let file = tempfile().unwrap();
 /// let rights = FcntlsBuilder::new()
-///     .add(Fcntl::GetFL)
+///     .allow(Fcntl::GetFL)
 ///     .finalize();
 ///
 /// rights.limit(&file).unwrap();